update runs: add extraction of generic package/mirror metadata
currently extracted: the "suite" metadata to map source packages to the Debian
suites they are present in.

As part of this change get rid of the "do nothing if the source package list is
unchanged". Rationale: it added a lot of complexity to udpate runs, made it
hard to add generic metadata extraction, and had near 0% hit rate.