add new SUITES field to cache/sources.txt
Semantics: comma separate list of suites a given source package is
currently part of.

As this is the first case of update information that crosses update
phase boundaries, introduce a new StatusUpdate object, that is updated
throughout update phases. To avoid udpate races, anticipate the suite
update phases as 2nd phase, between exctract_new and garbage_collection.
It is that phase which is now in charge of saving sources.txt on disk.

Introduce new test updaterProducesReferenceSourcesTxt in
test_updater.py, to ensure that the udpater produces the expected
sources.txt file.