ctags/checksum plugins: do not use sqlalchmey ORM for inserts
Rather, use prepared SQL statements to do bulk inserts. Significant time saving
(about 2x speedup) which actually makes the DB structure with a factored out
File table faster (~40%) than the pre-refactoring structure.

Note that even if we do not use the ORM we still use the same session, as we
want to be transactional with the rest of the update.

This change also improves memory efficient, but on that front it is still too
memory-heavy, e.g. injection of chromium's ctags takes ~5-6 GB, which is
definitely too much.