allow redirects to package versions based on suite/codename
We now check for suite names to redirect to a specific version. If package foo
has version 1.5-1 in sid, going to /src/foo/sid will display as if we went to
/src/foo/1.5-1.

How this works: If we go to /src/foo/sid for example, we iterate through the
package's (foo) suites. If we find 'sid', we simply handle the request as if it
was /src/foo/1.5-1.

This commit also adds a "suites_aliases" table which let us setup aliases to
suites (ie: unstable->sid, testing->jessie). updater.py has also been updated
to auto-generate those aliases, by parsing the symlinks in the dists/ folders.

When a user requests /src/foo/unstable, it will lookup the suites_aliases table
and check if this alias is mapped to a suite. If yes, it will handle the
request as if the user requested that particular suite (in this case,
/src/foo/sid). Which in turns would handle it as if it was /src/foo1.5-1.
(taking the example above)

Signed-off-by: Stefano Zacchiroli <zack@upsilon.cc>