web app: fix 'undefined' line numbers in source code pages
Problem: In Iceweasel / Firefox, clicking on a line-number when browsing
a source file caused the line number returned to be "undefined", and
so "#Lundefined" appeared in the URL.

The cause was that to retrieve the line numbers from the <a> elements we
clicked on, I was using .innerText, which is not supported by Firefox /
Iceweasel. The alternative is using .textContent.

The solution was to use "element.textContent || element.innerText", so
that it works browsers that support either .textContent or .innerText.

Closes: #774808

Signed-off-by: Stefano Zacchiroli <zack@upsilon.cc>