source_folder: add a show/hide hidden files button
Since that hidden files and directories are still passed to the view, but are
now given an 'hidden' boolean attribute, we can show or hide them once the page
is loaded.

This patch also changes the way javascript is loaded by debsources.js, so that
it does not run unnecessary code (for example we don't need to print line
numbers in folders).

Signed-off-by: Stefano Zacchiroli <zack@upsilon.cc>