test suite: add @notravis attribute, apply it to refdb test
Rationale: the @notravis attribute will be temporarily used to tag all tests
that do not run properly in travis test environment. They still are test smells
that should be fixed, see bug #783832. But in the meantime this approach is
better than excluding the test for running for everyone.

Apply the @notravis attribute to test_updater.producesReferenceDb test, which
currently fails within travis. See for instance:
https://travis-ci.org/Debian/debsources/builds/61977243

Add a "test-travis" target to Makefile, which will run all tests that are not
tagged with @notravis.

Amend .travis.yml to make it run "make test-travis".