Docker: split in two images, and use docker-compose
Instead of running one container running postgres, apache and debsources
(flask application), we now have a container running postgres, and
another one running the web app.

This allows us to have a container that holds data (postgres), that we'd
like to keep around. The webapp container can be recreated at will,
since the code itself is on the host's drive and shared as a volume
inside the container.

docker-compose.yml has to be in the project's root for now
(unfortulately). It runs docker with the correct port mappings, volumes,
and links our two containers correctly.

 * To build images: bin/debsources-docker-build
 * To run the containers: bin/debsources-docker-run (this will start both
   postgres and the webapp)
 * To update the postgres data: bin/debsources-docker-update