fix: order of config files which are looked for
config.local.ini should be used in priority, compared to /etc or /srv,
to allow many instances of debsources to be run on the same machine.

this won't break actual configurations, except if a config.local.ini
exists along with one in /etc. This is not the case in Docker, nor in
the recommended development setups in the doc.