stats web page: comment out 20-year graphs for now
We do not have the data to fill them right now, they are just sampled down
equivalent of the 5-year graphs, and they do puzzle people all the time. We can
reenable them when we have more data. Note that the plots are still generated,
they are just not shown in the HTML template.