DB schema: upgrade most integers types to bigint
To avoid data errors with large integers, as it is starting to happen for the
ctags.id column. The only left integer (not bigint) columns are stuff like line
position in the ctags table, and counters for generally small quantities, e.g.,
source/binary package amounts.

Commit includes both changes to models.py and the SQL upgrade script.