debmirror launcher: use --ignore-missing-release
Some virtual suites, like oldoldstable and friends, come and go from the
mirror network. By default, debmirror fails if any of the suites given
to it are missing from the mirror network. This change allows debmirror
to continue, even when oldoldstable is missing, and relieves us from the
obligation of constantly having to fiddle with the list of suites listed
in the configuration file.