webapp: Add support for old news
Closes: #783911 - "old news page"

Previously only news option has been supported in the web application.
This patch introduces support for old news as well. The old news file
is added manually and it results in a link on the main page which points
to the old news file. This is available for all the blueprints. The file
name should be _old_news.html for the blueprints where _news.html is used
previously for the news file.

Signed-off-by: Orestis Ioannou <orestis@oioannou.com>