Pull-up common testdata/ instructions
The instructions for fetching testdata/ are also needed for the Docker
build. As a newcomer, I just followed the Docker build instructions step
and faced the following error after running "make update-db":

  No such file or directory: '/opt/debsources/testdata/mirror/dists'

Making it clear that this step also needs to be run for the Docker build
will help other newcomers to not trip over the same stone.