Remove Flask app singleton
For practical purposes, a singleton (via a module variable) was used
to store a configured Flask application. When used in testing, this
object was configured over and over. This is now failing with a recent
version of Flask, which does not accept redefining exising blueprints.

Remove singleton altogether, following Flask's app factory pattern.